Fix: use first nonce in the ExecuteTxButton + add tests
