+++ src/components/transactions/ExecuteTxButton/index.tsx
@@ -3,7 +3,6 @@ import { type ReactElement, useContext } from 'react'
 import { type TransactionSummary } from '@safe-global/safe-gateway-typescript-sdk'
 import { Button, Tooltip } from '@mui/material'
 
-import useSafeInfo from '@/hooks/useSafeInfo'
 import { isMultisigExecutionInfo } from '@/utils/transaction-guards'
 import Track from '@/components/common/Track'
 import { TX_LIST_EVENTS } from '@/services/analytics/events/txList'
@@ -12,6 +11,7 @@ import CheckWallet from '@/components/common/CheckWallet'
 import { useSafeSDK } from '@/hooks/coreSDK/safeCoreSDK'
 import { TxModalContext } from '@/components/tx-flow'
 import { ConfirmTxFlow } from '@/components/tx-flow/flows'
+import { useFirstQueuedNonce } from '@/hooks/useTxQueue'
 
 const ExecuteTxButton = ({
   txSummary,
@@ -21,12 +21,12 @@ const ExecuteTxButton = ({
   compact?: boolean
 }): ReactElement => {
   const { setTxFlow } = useContext(TxModalContext)
-  const { safe } = useSafeInfo()
   const txNonce = isMultisigExecutionInfo(txSummary.executionInfo) ? txSummary.executionInfo.nonce : undefined
   const { setSelectedTxId } = useContext(ReplaceTxHoverContext)
   const safeSDK = useSafeSDK()
+  const firstNonce = useFirstQueuedNonce()
 
-  const isNext = txNonce !== undefined && txNonce === safe.nonce
+  const isNext = txNonce !== undefined && txNonce === firstNonce
   const isDisabled = !isNext || !safeSDK
 
   const onClick = (e: SyntheticEvent) => {
