+++ src/components/transactions/GroupLabel/index.tsx
@@ -1,8 +1,13 @@
 import type { ReactElement } from 'react'
 import type { Label } from '@safe-global/safe-gateway-typescript-sdk'
 import { LabelValue } from '@safe-global/safe-gateway-typescript-sdk'
-import QueueLabelText from './QueueLabelText'
+import { useFirstQueuedNonce } from '@/hooks/useTxQueue'
 import css from './styles.module.css'
+
+const QueueLabelText = () => {
+  const firstNonce = useFirstQueuedNonce()
+  return ` - transaction with nonce ${firstNonce} needs to be executed first`
+}
 
 const GroupLabel = ({ item }: { item: Label }): ReactElement => {
   return (
