+++ src/hooks/__tests__/useTxQueue.test.ts
@@ -1,8 +1,10 @@
-import { useQueuedTxsLength } from '../useTxQueue'
+import { useFirstQueuedNonce, useQueuedTxsLength } from '../useTxQueue'
 import * as store from '@/store'
-import * as recoveryHooks from '../../features/recovery/hooks/useRecoveryQueue'
+import * as recoveryHooks from '@/features/recovery/hooks/useRecoveryQueue'
+import * as safeInfoHooks from '../useSafeInfo'
 import type { RecoveryQueueItem } from '@/features/recovery/services/recovery-state'
-import { TransactionListItemType } from '@safe-global/safe-gateway-typescript-sdk'
+import { DetailedExecutionInfoType, TransactionListItemType } from '@safe-global/safe-gateway-typescript-sdk'
+import { renderHook } from '@/tests/test-utils'
 
 describe('useQueuedTxsLength', () => {
   beforeEach(() => {
@@ -71,5 +73,79 @@ describe('useQueuedTxsLength', () => {
 
     const result = useQueuedTxsLength()
     expect(result).toEqual('5')
+  })
+
+  describe('useFirstQueuedNonce', () => {
+    it('should return the nonce of the first queued transaction', () => {
+      jest.spyOn(safeInfoHooks, 'default').mockReturnValue({ safeAddress: '0x1234', safe: { nonce: -1 } } as any)
+
+      jest.spyOn(store, 'useAppSelector').mockReturnValue({
+        data: {
+          results: [
+            {
+              type: TransactionListItemType.LABEL,
+            },
+            {
+              type: TransactionListItemType.TRANSACTION,
+              transaction: {
+                executionInfo: {
+                  type: DetailedExecutionInfoType.MODULE,
+                },
+              },
+            },
+            {
+              type: TransactionListItemType.TRANSACTION,
+              transaction: {
+                executionInfo: {
+                  type: DetailedExecutionInfoType.MULTISIG,
+                  nonce: 1,
+                },
+              },
+            },
+            {
+              type: TransactionListItemType.TRANSACTION,
+              transaction: {
+                executionInfo: {
+                  type: DetailedExecutionInfoType.MULTISIG,
+                  nonce: 2,
+                },
+              },
+            },
+          ],
+
+          next: undefined,
+        },
+      })
+
+      const { result } = renderHook(() => useFirstQueuedNonce())
+      expect(result.current).toEqual(1)
+    })
+
+    it('should return the safe nonce if no queued transactions', () => {
+      jest.spyOn(safeInfoHooks, 'default').mockReturnValue({ safeAddress: '0x1234', safe: { nonce: 10 } } as any)
+
+      jest.spyOn(store, 'useAppSelector').mockReturnValue({
+        data: {
+          results: [
+            {
+              type: TransactionListItemType.LABEL,
+            },
+            {
+              type: TransactionListItemType.TRANSACTION,
+              transaction: {
+                executionInfo: {
+                  type: DetailedExecutionInfoType.MODULE,
+                },
+              },
+            },
+          ],
+
+          next: undefined,
+        },
+      })
+
+      const { result } = renderHook(() => useFirstQueuedNonce())
+      expect(result.current).toEqual(10)
+    })
   })
 })
